chore: move wapc host callbacks to its own mod
